fix: type error of DEFAULT_CLAIM_BLOCKS

+++ src/main/java/me/drex/itsours/util/Constants.java
@@ -6,7 +6,7 @@ import net.minecraft.item.Items;
 // TODO: Make configurable
 public class Constants {
 
-    public static final int DEFAULT_CLAIM_BLOCKS = 50000;
+    public static final long DEFAULT_CLAIM_BLOCKS = 50000;
     public static final Item SELECT_ITEM = Items.GOLDEN_SHOVEL;
 
     public static final int DEFAULT_CLAIM_COUNT = 5;
